multithreading implemented and heic support fixed

+++ loffPDF.py
@@ -4,19 +4,31 @@ import tkinter.filedialog
 from tkinter import messagebox
 import ntpath
 from wand.image import Image
-import threading
 from pathlib import Path
+from concurrent.futures import ProcessPoolExecutor
+from PIL import Image as PILImage
+import pillow_heif
+
+supported_savetypes = [
+    ".pdf",
+    ".jpg",
+    ".png"
+]
 
 supported_filetypes = [
     ".pdf",
     ".jpg",
     ".png",
-    ".tiff",
     ".heic",
-    ".bmp"
+    "tiff",
+    "bmp"
 ]
 
 status_out = ""
+
+# MULTIPROCESSING
+exe = ProcessPoolExecutor()
+futures = []
 
 
 class App(tk.Frame):
@@ -47,13 +59,13 @@ class App(tk.Frame):
         self.bt_flatten = ttk.Button(self, text="Flatten/Fix PDFs", command=self.do_flatten)
         self.bt_flatten.grid(column=0, row=2)
 
-        self.bt_flatten = ttk.Button(self, text="Convert to new format:", command=self.do_convert)
-        self.bt_flatten.grid(column=1, row=2)
+        self.bt_convert = ttk.Button(self, text="Convert to new format:", command=self.do_convert)
+        self.bt_convert.grid(column=1, row=2)
 
         self.dd_var = tk.StringVar()
-        self.dd_var.set(supported_filetypes[0])
+        self.dd_var.set(supported_savetypes[0])
 
-        self.dd_filetypes = ttk.OptionMenu(self, self.dd_var, *supported_filetypes)
+        self.dd_filetypes = ttk.OptionMenu(self, self.dd_var, *supported_savetypes)
         self.dd_filetypes.grid(column=2, row=2)
 
         self.tx_output = tk.Text(self, height=8)
@@ -83,26 +95,68 @@ class App(tk.Frame):
         self.open_file(True)
 
     def do_flatten(self, *args):
-        self.output_write("", clear=True)
-        run = True
-        for file in self.file_list:
-            if ".pdf" not in file:
-                run = False
-                messagebox.showerror("Check files!", "Flatten only supports PDF files, please remove other filetypes.")
-                break
-        
-        if run:
-            flatten(self.file_list)
-            self.lb_file_list.delete(0,tk.END)
-            self.file_list.clear()
+        if len(self.file_list) != 0:
+            self.output_write("", clear=True)
+            self.output_write("Flattening " + str(len(self.file_list)) + " files")
+
+            run = True
+            for file in self.file_list:
+                if ".pdf" not in file:
+                    run = False
+                    messagebox.showerror("Check files!", "Flatten only supports PDF files, please remove other filetypes.")
+                    break
+            
+            if run:
+                for file in self.file_list:
+                    future = exe.submit(flatten, file)
+                    future.add_done_callback(self.task_callback)
+                    futures.append(future)
+
+                self.lb_file_list.delete(0,tk.END)
+                self.file_list.clear()
+                self.bt_flatten.config(state="disabled")
+                self.bt_convert.config(state="disabled")
+                self.pb_progress.config(mode="indeterminate")
+                self.pb_progress.start()
 
     def do_convert(self, *args):
-        self.output_write("", clear=True)
-        convert(self.file_list, self.dd_var.get())
-        self.lb_file_list.delete(0,tk.END)
-        self.file_list.clear()
+        if len(self.file_list) != 0:
+            self.output_write("", clear=True)
+            self.output_write("Converting " + str(len(self.file_list)) + " files")
 
 
+            for file in self.file_list:
+                    future = exe.submit(convert, file, self.dd_var.get())
+                    future.add_done_callback(self.task_callback)
+                    futures.append(future)
+
+            self.lb_file_list.delete(0,tk.END)
+            self.file_list.clear()
+            self.bt_convert.config(state="disabled")
+            self.bt_flatten.config(state="disabled")
+            self.pb_progress.config(mode="indeterminate")
+            self.pb_progress.start()
+
+    def reset_gui(self):
+        self.bt_flatten.config(state="normal")
+        self.bt_convert.config(state="normal")
+        self.pb_progress.config(mode="determinate")
+        self.pb_progress.stop()
+        self.pb_progress.step(99.9)
+
+    def task_callback(self, future):
+        if future.exception() is None:
+            self.output_write(future.result())
+            if all(task.done() for task in futures):
+                futures.clear()
+                self.output_write("Complete.")
+                self.reset_gui()
+            else:
+                self.pb_progress.config(mode="determinate")
+                self.pb_progress.stop()
+                self.pb_progress.step((sum(task.done() for task in futures) / len(futures)) * 99.9)
+        else:
+            self.output_write("An error occured")
 
     def remove_item(self, *args):
         selection = self.lb_file_list.curselection()
@@ -128,6 +182,36 @@ class App(tk.Frame):
         self.tx_output.config(state="disabled")
 
 
+
+# WORKER TASKS
+def flatten(file, *options):
+    with Image(filename=file) as src:
+        with src.convert("pdf") as out:
+            output = Path(file).stem + "_flattened.pdf"
+            out.save(filename=output)
+
+    return "Flattened: " + output
+
+def convert(file, filetype, *options):
+        if Path(file).suffix != ".heic":
+            with Image(filename=file) as src:
+                with src.convert(filetype.replace(".", "")) as out:
+                    output = Path(file).stem + "_converted" + filetype
+                    out.save(filename=output)
+        else:
+            raw = pillow_heif.read_heif(file)
+            src = PILImage.frombytes(raw.mode,
+                raw.size,
+                raw.data,
+                "raw",)
+            output = Path(file).stem + "_converted" + filetype
+            if filetype == ".jpg":
+                src.convert("RGB").save(output)
+            else:
+                src.save(output)
+
+        return "Converted: " + output
+
 # MAIN APP
 
 root = tk.Tk()
@@ -140,29 +224,5 @@ root.rowconfigure(0, weight=1)
 
 app = App(root)
 app.grid(column=0, row=0, sticky="nsew", padx=4, pady=4)
-        
-
-#TODO THREADPOOL WITH CALLBACKS
-
-
-# WORKER TASKS
-def flatten(files, *options):
-    for filePath in files:
-        with Image(filename=filePath) as src:
-            with src.convert("pdf") as out:
-                output = Path(filePath).stem + "_flattened.pdf"
-                out.save(filename=output)
-
-        app.output_write("Flatten done: " + output)
-
-def convert(files, filetype, *options):
-    for filePath in files:
-        with Image(filename=filePath) as src:
-            with src.convert(filetype.replace(".", "")) as out:
-                output = Path(filePath).stem + "_converted" + filetype
-                out.save(filename=output)
-
-        app.output_write("Convert done: " + output)
-
 
 root.mainloop()
